# Optimize rational comparison