It's a change to `do_compare`, which becomes faster and conses less. This is what we do: instead of doing (x-y) and checking if the result is negative, positive or zero, when one of the arguments is rational, we cross-multiply and call `do_compare` recursively.

Comment from the source:

```
when comparing a rational to integers, or to another rational,
instead of computing the difference, we compare
a/b compared to c/d  <-->  ad compared to bc
a   compared to c/d  <-->  ad compared to c
and so on. This is faster and conses less.
But we do NOT do this for reals and complexes, because it's a
more complicated situation.
```

In particular, when there are only fixnums involved (in the denominators/numerators etc) we don't cons at all, and the speedup is really big.

Results:

```scheme
;; BIGNUM-RAT x BIGNUM-RAT
(let ((a (/ (expt 2 300)
            (expt 5 100)))
      (b (/ (expt 3 300)
            (expt 5 100))))
  (time
   (repeat 1_000_000
           (< a b))))
;; 3061.321 ms
;; 1104000000 bytes in 22000000 allocation calls
;; 965.31 ms
;; 616000000 bytes in 10000000 allocation calls

;; FIXNUM-RAT x FIXNUM-RAT
(let ((a (/ (expt 2 15)
            (expt 5 10)))
      (b (/ (expt 3 15)
            (expt 5 10))))
  (time
   (repeat 10_000_000
           (< a b))))
;; 4606.831 ms
;; 240000000 bytes in 10000000 allocation calls
;; 940.646 ms
;; 0 bytes in 0 allocation calls

;; FIXNUM-RAT x BIGNUM-RAT
(let ((a (/ (expt 2 15)
            (expt 5 10)))
      (b (/ (expt 3 300)
            (expt 5 100))))
  (time
   (repeat 1_000_000
           (< a b))))
;; 2711.321 ms
;; 808000000 bytes in 23000000 allocation calls
;; 1001.095 ms
;; 432000000 bytes in 11000000 allocation calls

;; FIXNUM x FIXNUM-RAT
(let ((a (expt 5 10))
      (b (/ (expt 3 15)
            (expt 5 10))))
      (time
       (repeat 10_000_000
               (< a b))))
;; 3998.837 ms
;; 240000000 bytes in 10000000 allocation calls
;; 735.415 ms
;; 0 bytes in 0 allocation calls

;; FIXNUM x BIGNUM-RAT
(let ((a (expt 5 10))
      (b (/ (expt 3 300)
            (expt 5 10))))
      (time
       (repeat 5_000_000
               (< a b))))
;; 2781.935 ms
;; 1000000000 bytes in 30000000 allocation calls
;; 1770.585 ms
;; 840000000 bytes in 20000000 allocation calls

;; BIGNUM x FIXNUM-RAT
(let ((a (expt 5 200))
      (b (/ (expt 3 15)
            (expt 5 10))))
      (time
       (repeat 1_000_000
               (< a b))))
;; 838.204 ms
;; 368000000 bytes in 10000000 allocation calls (GC: 94) ;; 644.563 ms
;; 336000000 bytes in 8000000 allocation calls (GC: 82)

;; BIGNUM x BIGNUM-RAT
(let ((a (expt 5 200))
      (b (/ (expt 3 300)
            (expt 5 10))))
      (time
       (repeat 1_000_000
               (< a b))))
;; 754.978 ms
;; 360000000 bytes in 9000000 allocation calls (GC: 94) 
;; 596.94 ms
;; 328000000 bytes in 7000000 allocation calls (GC: 82)
```