@jpellegrini 2025-05-24
Although a solution to #769 would perhaps make this PR unnecessary?