@egallesio 2025-05-25
That is very smart! I

> Although a solution to #769 would perhaps make this PR unnecessary?

Yes, but it will always be more efficient than allocating, no?