@jpellegrini 2025-05-25
> Yes, but it will always be more efficient than allocating, no?

You're right... :)  So maybe this one can me merged?