@jpellegrini 2025-07-10
Added some more optimizations:


1. When comparing pos/neg or neg/pos, immediately answer instead of
   computing a-b.
2. When comparing a bignum and a fixnum, the sign of the bignum alone
   determines the result, regardless of the values of both numbers.

```scheme
(let ((a 10)
      (b (- (expt 3 700))))
  (time (repeat 6_000_000 (> a b))))

Old code: 1732.233 ms, 1968000000 bytes in 24000000 allocations
New code:  198.913 ms,          0 bytes in        0 allocations

(let ((a (expt 5 200))
      (b (- (expt 3 700))))
  (time (repeat 6_000_000 (> a b))))

1519.886 ms, 1920000000 bytes in 18000000 allocations
 233.08  ms,          0 bytes in        0 allocations
```