@jpellegrini 2025-07-10
The second commit also uncovered a bug in `(scheme bytevector)` (Issue #792 ) -- that's why two tests fail after this PR.